ref: breaking pkg for better API usage. Adding P2P endpoints
