+++ examples/txs/main.go
@@ -21,13 +21,14 @@ func init() {
 }
 
 func main() {
-	api := qvapay.NewPaymentAppClient(
-		os.Getenv("APP_ID"),     // app_id
-		os.Getenv("APP_SECRET"), // secret_id
-		qvapay.BaseURL,          // constants url base https://qvapay.com/api
-		false,                   // skip verificationSSL
-		nil,                     // custom http.PaymentAppClient
-		nil,                     // debug io.Writter (os.Stdout)
+	api := qvapay.NewPaymentAppClient(qvapay.Options{
+		BaseURL:    qvapay.BaseURL,          // constants url base https://qvapay.com/api
+		HttpClient: nil,                     // custom http.PaymentAppClient
+		Debug:      os.Stdout,               // debug io.Writter (os.Stdout)
+		AppID:      os.Getenv("APP_ID"),     // app_id
+		SecretID:   os.Getenv("APP_SECRET"), // secret_id
+		SkipVerify: false,                   // skip verificationSSL
+	},
 	)
 
 	tx, err := api.GetTransactions(
