--- http_uitl.go
@@ -1,77 +1,0 @@
-package qvapay
-
-import (
-	"bytes"
-	"context"
-	"fmt"
-	"io"
-	"io/ioutil"
-	"log"
-	"net/http"
-	"net/http/httputil"
-)
-
-// dumpResponse writes the raw response data to the debug output, if set, or
-// standard error otherwise.
-func (c *client) dumpResponse(resp *http.Response) {
-	// ignore errors dumping response - no recovery from this
-	responseDump, err := httputil.DumpResponse(resp, true)
-	if err != nil {
-		log.Fatalf("dumpResponse: " + err.Error())
-	}
-	fmt.Fprintln(c.debug, string(responseDump))
-	fmt.Fprintln(c.debug)
-}
-
-// apiCall define how you can make a call to qvapay API
-func (c *client) apiCall(
-	ctx context.Context,
-	method string,
-	URL string,
-	data []byte,
-) (statusCode int, response string, err error) {
-
-	req, err := http.NewRequest(method, URL, bytes.NewBuffer(data))
-	if err != nil {
-		return 0, "", fmt.Errorf("failed to create HTTP request: %v", err)
-	}
-	req.Header.Add("content-type", "application/json")
-	req.Header.Set("User-Agent", "qvapaygo-client/0.0")
-	if c.debug != nil {
-		requestDump, err := httputil.DumpRequestOut(req, true)
-		if err != nil {
-			return 0, "", fmt.Errorf("error dumping HTTP request: %v", err)
-		}
-		fmt.Fprintln(c.debug, string(requestDump))
-		fmt.Fprintln(c.debug)
-	}
-	req = req.WithContext(ctx)
-	resp, err := c.httpClient.Do(req)
-	if err != nil {
-		return 0, "", fmt.Errorf("HTTP request failed with: %v", err)
-	}
-	defer DrainBody(resp.Body)
-	if c.debug != nil {
-		c.dumpResponse(resp)
-	}
-	res, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return resp.StatusCode, "", fmt.Errorf("HTTP request failed: %v", err)
-	}
-	return resp.StatusCode, string(res), nil
-}
-
-func DrainBody(respBody io.ReadCloser) {
-	// Callers should close resp.Body when done reading from it.
-	// If resp.Body is not closed, the PaymentAppClient's underlying RoundTripper
-	// (typically Transport) may not be able to re-use a persistent TCP
-	// connection to the server for a subsequent "keep-alive" request.
-	if respBody != nil {
-		// Drain any remaining Body and then close the connection.
-		// Without this closing connection would disallow re-using
-		// the same connection for future uses.
-		//  - http://stackoverflow.com/a/17961593/4465767
-		defer respBody.Close()
-		_, _ = io.Copy(ioutil.Discard, respBody)
-	}
-}
